Couldn't get saving to work.
UpdateFragment.java
Added a plus on the toolbar which lets me add a new myQueueItem that creates a new assignment

sendToServerObject.java
Moved some funtions that create StringEntity from myQueue.java to here. All of the network traffic will be handled with this object. Can take a myQueueTask or a myQueueItem in a constructor and can be called using .send() to send to the respective URL

myExpandRecyclerAdapter.java
Set an onClickListened to the 'SEND' button which opens a yes or no dialog to send the data to the server. It calls sendObjectToServer(myQueueTask).send()

Everything else:
Small fixes / Optimizations